zephyr: Implement Dhara disk driver as Zephyr module
Zephyr has currently no flash translation layer implemented. This commit
adds the basic possibility to use Dhara as FTL in Zephyr. It is
implemented as disk driver that can be placed between a filesystem and a
flash driver. A symbolic link is created to fulfill the requirement that
include paths shall begin with the module name.

Signed-off-by: Pascal Linder <pascal.linder@zuehlke.com>